refactor: #29 BaseModal에 맞게 Seatpicker, confirm 모달 수정
