+++ src/components/index.tsx
@@ -9,11 +9,13 @@ import Textarea from '@/components/common/Textarea/Textarea';
 import InputField from '@/components/common/Input/Input';
 import ReviewStepLayout from '@/components/review/ReviewLayout';
 import TagSection from '@/components/review/TagSection';
-import Modal from '@/components/common/Modal/Modal';
+import Modal from '@/components/common/Modal/BaseModal';
 import Header from '@/components/common/Header/Header';
 import ImagePreviewItem from './common/ImagePreview/ImagePreviewItem';
 import SeatItem from '@/components/seat/SeatItem';
 import SeatMap from '@/components/seat/SeatMap';
+import SeatPickerModal from '@/components/common/Modal/SeatPickerModal';
+import ConfirmModal from '@/components/common/Modal/ConfirmModal';
 
 export {
   InputField,
@@ -32,4 +34,6 @@ export {
   ImagePreviewItem,
   SeatItem,
   SeatMap,
+  SeatPickerModal,
+  ConfirmModal,
 };
